refactor(shared): centralize assistant identity and usage timeseries types
(cherry picked from commit 693f61404ddc88d9479a1524bb4b43fbc418d5d9)
