+++ src/infra/session-cost-usage.types.ts
@@ -1,4 +1,8 @@
 import type { NormalizedUsage } from "../agents/usage.js";
+import type {
+  SessionUsageTimePoint as SharedSessionUsageTimePoint,
+  SessionUsageTimeSeries as SharedSessionUsageTimeSeries,
+} from "../shared/session-usage-timeseries-types.js";
 
 export type CostBreakdown = {
   total?: number;
@@ -141,22 +145,9 @@ export type DiscoveredSession = {
   firstUserMessage?: string;
 };
 
-export type SessionUsageTimePoint = {
-  timestamp: number;
-  input: number;
-  output: number;
-  cacheRead: number;
-  cacheWrite: number;
-  totalTokens: number;
-  cost: number;
-  cumulativeTokens: number;
-  cumulativeCost: number;
-};
+export type SessionUsageTimePoint = SharedSessionUsageTimePoint;
 
-export type SessionUsageTimeSeries = {
-  sessionId?: string;
-  points: SessionUsageTimePoint[];
-};
+export type SessionUsageTimeSeries = SharedSessionUsageTimeSeries;
 
 export type SessionLogEntry = {
   timestamp: number;
